mm: thp: use single linked list for THP page table page deposit.
The old design uses the double linked list page->lru to chain all
deposited page table pages when creating a THP and page->pmd_huge_pte
to point to the first page of the list. As the second pointer in
page->lru overlaps with page->pmd_huge_pte, the design prevents
multi-level page table page deposit, which is useful for PUD and higher
level THPs.

The new design uses single linked list, where the deposit_head of the
THP pgtable points to a single linked list of deposited pgtable pages
chained with deposit_node. In addition, the deposit_node of the THP
pgtable can be used to form another list of deposited pgtable pages.
This allows us to have one PUD page points to a list of PMD pages, each
of which points a list of PTE pages to support PUD level THP.

__llist_del_first, a lock-protected version of llist_del_first, is added
to avoid expensive and unnecessary lockless operations, since page table
deposit is protected by page table locks.

Signed-off-by: Zi Yan <ziy@nvidia.com>