mm: truncate: split thp to a non-zero order if possible.
To minimize the number of pages after a truncation, when truncating a
THP, we do not need to split it all the way down to order-0. The THP has
at most three parts, the part before offset, the part to be truncated,
the part remaining at the end. Find the greatest common power of two
multiplier of the non-zero values of them as the new order. So we can
truncate a subpage as large as possible.

Signed-off-by: Zi Yan <ziy@nvidia.com>