mm: thp: add PUDDoubleMap page flag for PUD- and PMD-mapped pages.
A PUD THP can be mapped by PUD, PMD, and PTE entries. In addition to
existing DoubleMap, which is for PMD- and PTE-mapped pages, we need
a new page flag for PUD- and PMD-mapped pages. PUDDoubleMap uses the
same page flag bit as DoubleMap but sets it at the second tail page
instead of the first tail page.

Signed-off-by: Zi Yan <ziy@nvidia.com>