mm: Make page_index() THP-safe
You can't dereference page->index for tail pages, so use the head's
index to calculate the offset of this page in the file.  Also move
the function to pagemap.h since it's a pagecache function not used
by the rest of the kernel.  This also makes page_file_offset() work
when called for tail pages.

Signed-off-by: Matthew Wilcox (Oracle) <willy@infradead.org>